cleaned up structure, adjusted fonts and paddings, simplified some of the code, added new buttons and removed overlay
