+++ packages/augur-ui/src/assets/styles/_misc_redesign.less
@@ -51,7 +51,6 @@
 
 // default measurements
 @button-height-default: @size-40; // 40px
-@button-padding-orders: @size-20 5.875rem; // 20px x 94px
 
 // borders
 @border-default: 1px solid transparent;
