+++ packages/augur-ui/src/modules/trading/components/form.styles.less
@@ -19,7 +19,7 @@
         min-height: @size-32;
 
         > span {
-          .mono-10;
+          .mono-10-bold;
 
           max-height: @size-32;
           min-height: @size-32;
@@ -58,7 +58,7 @@
         min-height: @size-16;
 
         > button {
-          .mono-10;
+          .mono-10-bold;
         }
 
         > div {
@@ -66,7 +66,7 @@
           flex: 1;
 
           > button {
-            .mono-10;
+            .mono-10-bold;
           }
 
           > button:not(:first-of-type) {
@@ -104,7 +104,7 @@
         width: fit-content;
 
         > button {
-          .mono-10;
+          .mono-10-bold;
         }
       }
 
@@ -120,6 +120,10 @@
         }
 
         > div:first-of-type {
+          > button {
+            .mono-10-bold;
+          }
+          
           height: fit-content;
         }
