+++ packages/augur-ui/src/modules/trading/components/form.tsx
@@ -10,7 +10,7 @@ import {
   SELL,
 } from 'modules/common/constants';
 import FormStyles from 'modules/common/form-styles.less';
-import Styles from 'modules/trading/components/form/form.styles.less';
+import Styles from 'modules/trading/components/form.styles.less';
 import { ExclamationCircle } from 'modules/common/icons';
 import { SquareDropdown } from 'modules/common/selection';
 import { Checkbox, TextInput } from 'modules/common/form';
